Tracing controller charm
Adds tracing to the controller charm. This is very crude, and we
should codify it into the ops framework if we want to actually use
it. Or at the very least implement it as a subclassed
TracedCharmBase, so it becomes opt-in.

The basic concept is there, it would be nice to get this fleshed out
at some point.