Always update relation settings on relation changed.
The controller charm attempted to minimise rewriting relation settings
by maintaining all_db_bind_addresses and only when it detected
differences.

This is not necessary, because as long as the same data has the same
representation (JSON in the application data bag for example), this
doesn’t cause settings watchers to fire.

Now we always just rewrite state.